test: split agent claim snapshot coverage
